VottingApp/requirement.txt
Django==1.8
MySQL-python==1.2.5
django-tastypie==0.13.3
python-dateutil==2.5.2
python-mimeparse==1.5.1
six==1.10.0
wsgiref==0.1.2
